MPGArray: constructor

MultiPrecisionArrays.MPGArray — Method

MPGArray(AH::AbstractArray{Float64,2}; basissize=10, TL=Float32)

An MPGArray stores the high precision matrix, the low precision factorization, the Krylov basis, and a few other things GMRES needs. If the high precision matrix is double, the low precision is single by default. Half is an optioin which you get with TL=Float16.

MultiPrecisionArrays.MPGArray — Method

MPGArray(AH::AbstractArray{Float32,2}; basissize=10, TL=Float16)

An MPGArray stores the high precision matrix, the low precision factorization, the Krylov basis, and a few other things GMRES needs. Since High precision is single, low is half. I'm leaving the kwarg for TL in there because it makes is easier to cut/paste calls to MPGArray different precisions into a CI loop.