mpgeslir: IR solver

MultiPrecisionArrays.mpgeslir — Method

mpgeslir(AF::MPFact, b; reporting=false, verbose=false, termparms=termparms_default)

I do not export this function. The idea is that you use mplu and do not touch either the constructor or the solver directly.

Use a multi-precision factorization to solve a linear system with plain vanilla iterative refinement.

This version is analogous to A\b and combines the factorization and the solve. You start with MPA=MPArray(A) and then pass MPA to mpgeslir and combine the factorization and the solve.

You can also get the multiprecision factorization directly with

MPF=mplu(A)

and then pass MPF to mpgeslir.

I use this to get some timing results and it's also convenient if you want to do factor and solve in one statement.

You can also get this with x = MPA\b.

If you set the kwarg reporting to true you can get the IR residual history. The output of

x = MPA\b

Use a multi-precision factorization to solve a linear system with plain vanilla iterative refinement.

MPFact is a union of all the MultiPrecision factorizations in the package. The triangular solver will dispatch on the various types depending on how the interprecision transfers get done.

MultiPrecisionArrays.mpgeslir — Method

mpgeslir(MPA::MPArray, b; reporting = false, verbose = false, termparms=termparms_default)

I do not export this function. The idea is that you use mpglu and do not touch either the constructor or the solver directly.

Use a multi-precision factorization to solve a linear system with plain vanilla iterative refinement.

This version is analogous to A\b and combines the factorization and the solve. You start with MPA=MPArray(A) and then pass MPA to mpgeslir and combine the factorization and the solve.

You can also get the multiprecision factorization directly with

MPF=mplu(A)

and then pass MPF to mpgeslir.

I use this to get some timing results and it's also convenient if you want to do factor and solve in one statement.

You can also get this with x = MPA\b.

If you set the kwarg reporting to true you can get the IR residual history. The output of

x = MPA\b

or

x=MPF\b

is the solition. The output of

mout = \(MPA,b; reporting=true)

or

mout = \(MPF,b; reporting=true)

is a structure. mpout.sol is the solution. mpout.rhist is the residual history. mpout.dhist is the history of the norms of the corrections. mpout also contains the datatypes TW for high precision and TF for low precision.

You may not get exactly the same results for this example on different hardware, BLAS, versions of Julia or this package. I am still playing with the termination criteria and the iteration count could grow or shrink as I do that.

Example

julia> using MultiPrecisionArrays.Examplesjulia> N = 4096; A = I - 800.0 * Gmat(N); b = ones(N);julia> MPF = mplu(A);julia> mout=\(MPF, b; reporting=true);julia> mout.rhist7-element Vector{Float64}: 1.00000e+00 4.24311e-02 9.03500e-05 9.59277e-08 6.12346e-11 9.75675e-12 8.49343e-12# Stagnation after six IR iterationsjulia> mout.dhist6-element Vector{Float64}: 2.00420e+02 3.15198e-01 4.48500e-04 4.57912e-07 5.88713e-10 3.62408e-11# No correction for the final itertation.julia> [mout.TW mout.TF]1×2 Matrix{DataType}: Float64  Float32

mplu
mplu(A::AbstractArray{TW,2}; TF=nothing, TR=nothing, residterm=residtermdefault, onthefly=true) where TW <: Real