mpglu!: Factor a MPGArray and set it up for GMRES by allocating room

for Krylov vectors etc

MultiPrecisionArrays.mpglu! — Method

mpglu!(MPGA::MPGArray) Factor a MPGArray and set it up for GMRES by allocating room for Krylov vectors etc

This function factors the low precision copy and leaves the high precision matrix alone. mpglu! allocates storage for basissize Kylov vectors and some other things GMRES needs. You get a factorization object as output and can use \ to solve linear systems.