mplu: Combine MPArray construction and factorization

MultiPrecisionArrays.mplu — Method

mplu(A::AbstractArray{TW,2}; TF=nothing, onthefly=nothing) where TW <: Real

Combines the constructor of the multiprecision array with the factorization.

Step 1: build the MPArray

Step 2: factor the low precision copy and return the factorization object